Extract service setup to a separate file
