+++ src/vs/server/node/remoteExtensionHostAgentServer.ts
@@ -8,193 +8,33 @@ import * as fs from 'fs';
 import * as http from 'http';
 import * as net from 'net';
 import * as url from 'url';
-import { release, hostname } from 'os';
 import * as perf from 'vs/base/common/performance';
 import { performance } from 'perf_hooks';
 import { VSBuffer } from 'vs/base/common/buffer';
-import { Disposable, toDisposable } from 'vs/base/common/lifecycle';
+import { Disposable, DisposableStore } from 'vs/base/common/lifecycle';
 import { generateUuid } from 'vs/base/common/uuid';
-import { Promises } from 'vs/base/node/pfs';
 import { findFreePort } from 'vs/base/node/ports';
 import * as platform from 'vs/base/common/platform';
-import { PersistentProtocol, ProtocolConstants } from 'vs/base/parts/ipc/common/ipc.net';
+import { PersistentProtocol } from 'vs/base/parts/ipc/common/ipc.net';
 import { NodeSocket, WebSocketNodeSocket } from 'vs/base/parts/ipc/node/ipc.net';
 import { ConnectionType, ConnectionTypeRequest, ErrorMessage, HandshakeMessage, IRemoteExtensionHostStartParams, ITunnelConnectionStartParams, SignRequest } from 'vs/platform/remote/common/remoteAgentConnection';
 import { ExtensionHostConnection } from 'vs/server/node/extensionHostConnection';
 import { ManagementConnection } from 'vs/server/node/remoteExtensionManagement';
-import { createRemoteURITransformer } from 'vs/server/node/remoteUriTransformer';
-import { ILogService, LogLevel, AbstractLogger, DEFAULT_LOG_LEVEL, MultiplexLogService, getLogLevel, LogService } from 'vs/platform/log/common/log';
+import { ILogService } from 'vs/platform/log/common/log';
 import { FileAccess, Schemas } from 'vs/base/common/network';
-import product from 'vs/platform/product/common/product';
-import { IEnvironmentService, INativeEnvironmentService } from 'vs/platform/environment/common/environment';
-import { ServiceCollection } from 'vs/platform/instantiation/common/serviceCollection';
-import { SyncDescriptor } from 'vs/platform/instantiation/common/descriptors';
-import { ConfigurationService } from 'vs/platform/configuration/common/configurationService';
-import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { IRequestService } from 'vs/platform/request/common/request';
-import { RequestService } from 'vs/platform/request/node/requestService';
-import { ITelemetryAppender, NullAppender, supportsTelemetry } from 'vs/platform/telemetry/common/telemetryUtils';
-import { ITelemetryService, TelemetryLevel } from 'vs/platform/telemetry/common/telemetry';
-import { IExtensionGalleryService, IExtensionManagementCLIService, IExtensionManagementService } from 'vs/platform/extensionManagement/common/extensionManagement';
-import { ExtensionGalleryServiceWithNoStorageService } from 'vs/platform/extensionManagement/common/extensionGalleryService';
-import { ExtensionManagementService } from 'vs/platform/extensionManagement/node/extensionManagementService';
-import { InstantiationService } from 'vs/platform/instantiation/common/instantiationService';
-import { IDownloadService } from 'vs/platform/download/common/download';
-import { DownloadServiceChannelClient } from 'vs/platform/download/common/downloadIpc';
-import { ILocalizationsService } from 'vs/platform/localizations/common/localizations';
-import { LocalizationsService } from 'vs/platform/localizations/node/localizations';
-import { AppInsightsAppender } from 'vs/platform/telemetry/node/appInsightsAppender';
-import { ITelemetryServiceConfig } from 'vs/platform/telemetry/common/telemetryService';
-import { resolveCommonProperties } from 'vs/platform/telemetry/common/commonProperties';
-import { getMachineId } from 'vs/base/node/id';
-import { FileService } from 'vs/platform/files/common/fileService';
-import { DiskFileSystemProvider } from 'vs/platform/files/node/diskFileSystemProvider';
-import { IFileService } from 'vs/platform/files/common/files';
+import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { IProductService } from 'vs/platform/product/common/productService';
 import { RemoteAgentConnectionContext } from 'vs/platform/remote/common/remoteAgentEnvironment';
-import { IPCServer, ClientConnectionEvent, IMessagePassingProtocol, StaticRouter, ProxyChannel } from 'vs/base/parts/ipc/common/ipc';
-import { Emitter, Event } from 'vs/base/common/event';
-import { RemoteAgentEnvironmentChannel } from 'vs/server/node/remoteAgentEnvironmentImpl';
-import { RemoteAgentFileSystemProviderChannel } from 'vs/server/node/remoteFileSystemProviderServer';
-import { REMOTE_FILE_SYSTEM_CHANNEL_NAME } from 'vs/workbench/services/remote/common/remoteFileSystemProviderClient';
-import { RequestChannel } from 'vs/platform/request/common/requestIpc';
-import { ExtensionManagementChannel } from 'vs/platform/extensionManagement/common/extensionManagementIpc';
-import ErrorTelemetry from 'vs/platform/telemetry/node/errorTelemetry';
-import { ExtensionHostDebugBroadcastChannel } from 'vs/platform/debug/common/extensionHostDebugIpc';
-import { LogLevelChannel } from 'vs/platform/log/common/logIpc';
-import { IURITransformer } from 'vs/base/common/uriIpc';
 import { WebClientServer, serveError, serveFile } from 'vs/server/node/webClientServer';
 import { URI } from 'vs/base/common/uri';
 import { isEqualOrParent } from 'vs/base/common/extpath';
-import { IServerEnvironmentService, ServerEnvironmentService, ServerParsedArgs } from 'vs/server/node/serverEnvironmentService';
-import { basename, dirname, join } from 'vs/base/common/path';
-import { REMOTE_TERMINAL_CHANNEL_NAME } from 'vs/workbench/contrib/terminal/common/remoteTerminalChannel';
-import { RemoteTerminalChannel } from 'vs/server/node/remoteTerminalChannel';
+import { IServerEnvironmentService, ServerParsedArgs } from 'vs/server/node/serverEnvironmentService';
+import { dirname, join } from 'vs/base/common/path';
 import { LoaderStats } from 'vs/base/common/amd';
-import { RemoteExtensionLogFileName } from 'vs/workbench/services/remote/common/remoteAgentService';
-import { ExtensionManagementCLIService } from 'vs/platform/extensionManagement/common/extensionManagementCLIService';
-import { SpdLogLogger } from 'vs/platform/log/node/spdlogLog';
-import { IPtyService, TerminalSettingId } from 'vs/platform/terminal/common/terminal';
-import { PtyHostService } from 'vs/platform/terminal/node/ptyHostService';
-import { IRemoteTelemetryService, RemoteNullTelemetryService, RemoteTelemetryService } from 'vs/server/node/remoteTelemetryService';
-import { IUriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentity';
-import { UriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentityService';
-import { ICredentialsMainService } from 'vs/platform/credentials/common/credentials';
-import { CredentialsMainService } from 'vs/platform/credentials/node/credentialsMainService';
-import { EncryptionMainService } from 'vs/platform/encryption/node/encryptionMainService';
-import { RemoteTelemetryChannel } from 'vs/server/node/remoteTelemetryChannel';
 import { parseServerConnectionToken, ServerConnectionToken, ServerConnectionTokenParseError, ServerConnectionTokenType } from 'vs/server/node/serverConnectionToken';
-import { IEncryptionMainService } from 'vs/platform/encryption/common/encryptionService';
+import { setupServerServices, SocketServer } from 'vs/server/node/serverServices';
 
 const SHUTDOWN_TIMEOUT = 5 * 60 * 1000;
-
-const eventPrefix = 'monacoworkbench';
-
-class SocketServer<TContext = string> extends IPCServer<TContext> {
-
-	private _onDidConnectEmitter: Emitter<ClientConnectionEvent>;
-
-	constructor() {
-		const emitter = new Emitter<ClientConnectionEvent>();
-		super(emitter.event);
-		this._onDidConnectEmitter = emitter;
-	}
-
-	public acceptConnection(protocol: IMessagePassingProtocol, onDidClientDisconnect: Event<void>): void {
-		this._onDidConnectEmitter.fire({ protocol, onDidClientDisconnect });
-	}
-}
-
-function twodigits(n: number): string {
-	if (n < 10) {
-		return `0${n}`;
-	}
-	return String(n);
-}
-
-function now(): string {
-	const date = new Date();
-	return `${twodigits(date.getHours())}:${twodigits(date.getMinutes())}:${twodigits(date.getSeconds())}`;
-}
-
-class ServerLogService extends AbstractLogger implements ILogService {
-	_serviceBrand: undefined;
-	private useColors: boolean;
-
-	constructor(logLevel: LogLevel = DEFAULT_LOG_LEVEL) {
-		super();
-		this.setLevel(logLevel);
-		this.useColors = Boolean(process.stdout.isTTY);
-	}
-
-	trace(message: string, ...args: any[]): void {
-		if (this.getLevel() <= LogLevel.Trace) {
-			if (this.useColors) {
-				console.log(`\x1b[90m[${now()}]\x1b[0m`, message, ...args);
-			} else {
-				console.log(`[${now()}]`, message, ...args);
-			}
-		}
-	}
-
-	debug(message: string, ...args: any[]): void {
-		if (this.getLevel() <= LogLevel.Debug) {
-			if (this.useColors) {
-				console.log(`\x1b[90m[${now()}]\x1b[0m`, message, ...args);
-			} else {
-				console.log(`[${now()}]`, message, ...args);
-			}
-		}
-	}
-
-	info(message: string, ...args: any[]): void {
-		if (this.getLevel() <= LogLevel.Info) {
-			if (this.useColors) {
-				console.log(`\x1b[90m[${now()}]\x1b[0m`, message, ...args);
-			} else {
-				console.log(`[${now()}]`, message, ...args);
-			}
-		}
-	}
-
-	warn(message: string | Error, ...args: any[]): void {
-		if (this.getLevel() <= LogLevel.Warning) {
-			if (this.useColors) {
-				console.warn(`\x1b[93m[${now()}]\x1b[0m`, message, ...args);
-			} else {
-				console.warn(`[${now()}]`, message, ...args);
-			}
-		}
-	}
-
-	error(message: string, ...args: any[]): void {
-		if (this.getLevel() <= LogLevel.Error) {
-			if (this.useColors) {
-				console.error(`\x1b[91m[${now()}]\x1b[0m`, message, ...args);
-			} else {
-				console.error(`[${now()}]`, message, ...args);
-			}
-		}
-	}
-
-	critical(message: string, ...args: any[]): void {
-		if (this.getLevel() <= LogLevel.Critical) {
-			if (this.useColors) {
-				console.error(`\x1b[90m[${now()}]\x1b[0m`, message, ...args);
-			} else {
-				console.error(`[${now()}]`, message, ...args);
-			}
-		}
-	}
-
-	override dispose(): void {
-		// noop
-	}
-
-	flush(): void {
-		// noop
-	}
-}
 
 export type ServerListenOptions = { host?: string; port?: number; socketPath?: string };
 
@@ -214,9 +54,6 @@ declare module vsda {
 
 export class RemoteExtensionHostAgentServer extends Disposable {
 
-	private readonly _logService: ILogService;
-	private readonly _socketServer: SocketServer<RemoteAgentConnectionContext>;
-	private readonly _uriTransformerCache: { [remoteAuthority: string]: IURITransformer; };
 	private readonly _extHostConnections: { [reconnectionToken: string]: ExtensionHostConnection; };
 	private readonly _managementConnections: { [reconnectionToken: string]: ManagementConnection; };
 	private readonly _allReconnectionTokens: Set<string>;
@@ -225,25 +62,23 @@ export class RemoteExtensionHostAgentServer extends Disposable {
 	private shutdownTimer: NodeJS.Timer | undefined;
 
 	constructor(
-		private readonly _environmentService: IServerEnvironmentService,
-		private readonly _productService: IProductService,
+		private readonly _socketServer: SocketServer<RemoteAgentConnectionContext>,
 		private readonly _connectionToken: ServerConnectionToken,
 		hasWebClient: boolean,
-		REMOTE_DATA_FOLDER: string
+		REMOTE_DATA_FOLDER: string,
+		@IServerEnvironmentService private readonly _environmentService: IServerEnvironmentService,
+		@IProductService private readonly _productService: IProductService,
+		@ILogService private readonly _logService: ILogService,
 	) {
 		super();
 
-		const logService = getOrCreateSpdLogService(this._environmentService);
-		logService.trace(`Remote configuration data at ${REMOTE_DATA_FOLDER}`);
-		logService.trace('process arguments:', this._environmentService.args);
+		this._logService.trace(`Remote configuration data at ${REMOTE_DATA_FOLDER}`);
+		this._logService.trace('process arguments:', this._environmentService.args);
 		const serverGreeting = _productService.serverGreeting.join('\n');
 		if (serverGreeting) {
-			logService.info(`\n\n${serverGreeting}\n\n`);
+			this._logService.info(`\n\n${serverGreeting}\n\n`);
 		}
 
-		this._logService = new MultiplexLogService([new ServerLogService(getLogLevel(this._environmentService)), logService]);
-		this._socketServer = new SocketServer<RemoteAgentConnectionContext>();
-		this._uriTransformerCache = Object.create(null);
 		this._extHostConnections = Object.create(null);
 		this._managementConnections = Object.create(null);
 		this._allReconnectionTokens = new Set<string>();
@@ -254,141 +89,6 @@ export class RemoteExtensionHostAgentServer extends Disposable {
 			this._webClientServer = null;
 		}
 		this._logService.info(`Extension host agent started.`);
-	}
-
-	public async initialize(): Promise<{ telemetryService: ITelemetryService; }> {
-		const services = await this._createServices();
-		setTimeout(() => this._cleanupOlderLogs(this._environmentService.logsPath).then(null, err => this._logService.error(err)), 10000);
-		return services;
-	}
-
-	private async _createServices(): Promise<{ telemetryService: ITelemetryService; }> {
-		const services = new ServiceCollection();
-
-		// ExtensionHost Debug broadcast service
-		this._socketServer.registerChannel(ExtensionHostDebugBroadcastChannel.ChannelName, new ExtensionHostDebugBroadcastChannel());
-
-		// TODO: @Sandy @Joao need dynamic context based router
-		const router = new StaticRouter<RemoteAgentConnectionContext>(ctx => ctx.clientId === 'renderer');
-		this._socketServer.registerChannel('logger', new LogLevelChannel(this._logService));
-
-		services.set(IEnvironmentService, this._environmentService);
-		services.set(INativeEnvironmentService, this._environmentService);
-
-		services.set(ILogService, this._logService);
-		services.set(IProductService, this._productService);
-
-		// Files
-		const fileService = this._register(new FileService(this._logService));
-		services.set(IFileService, fileService);
-		fileService.registerProvider(Schemas.file, this._register(new DiskFileSystemProvider(this._logService)));
-
-		const configurationService = new ConfigurationService(this._environmentService.machineSettingsResource, fileService);
-		services.set(IConfigurationService, configurationService);
-
-		// URI Identity
-		services.set(IUriIdentityService, new UriIdentityService(fileService));
-
-		// Request
-		services.set(IRequestService, new SyncDescriptor(RequestService));
-
-		let appInsightsAppender: ITelemetryAppender = NullAppender;
-		const machineId = await getMachineId();
-		if (supportsTelemetry(this._productService, this._environmentService)) {
-			if (this._productService.aiConfig && this._productService.aiConfig.asimovKey) {
-				appInsightsAppender = new AppInsightsAppender(eventPrefix, null, this._productService.aiConfig.asimovKey);
-				this._register(toDisposable(() => appInsightsAppender!.flush())); // Ensure the AI appender is disposed so that it flushes remaining data
-			}
-
-			const config: ITelemetryServiceConfig = {
-				appenders: [appInsightsAppender],
-				commonProperties: resolveCommonProperties(fileService, release(), hostname(), process.arch, this._productService.commit, this._productService.version + '-remote', machineId, this._productService.msftInternalDomains, this._environmentService.installSourcePath, 'remoteAgent'),
-				piiPaths: [this._environmentService.appRoot]
-			};
-			const initialTelemetryLevelArg = this._environmentService.args['telemetry-level'];
-			let injectedTelemetryLevel: TelemetryLevel | undefined = undefined;
-			// Convert the passed in CLI argument into a telemetry level for the telemetry service
-			if (initialTelemetryLevelArg === 'all') {
-				injectedTelemetryLevel = TelemetryLevel.USAGE;
-			} else if (initialTelemetryLevelArg === 'error') {
-				injectedTelemetryLevel = TelemetryLevel.ERROR;
-			} else if (initialTelemetryLevelArg === 'crash') {
-				injectedTelemetryLevel = TelemetryLevel.CRASH;
-			} else if (initialTelemetryLevelArg !== undefined) {
-				injectedTelemetryLevel = TelemetryLevel.NONE;
-			}
-			services.set(IRemoteTelemetryService, new SyncDescriptor(RemoteTelemetryService, [config, injectedTelemetryLevel]));
-		} else {
-			services.set(IRemoteTelemetryService, RemoteNullTelemetryService);
-		}
-
-		services.set(IExtensionGalleryService, new SyncDescriptor(ExtensionGalleryServiceWithNoStorageService));
-
-		const downloadChannel = this._socketServer.getChannel('download', router);
-		services.set(IDownloadService, new DownloadServiceChannelClient(downloadChannel, () => this._getUriTransformer('renderer') /* TODO: @Sandy @Joao need dynamic context based router */));
-
-		services.set(IExtensionManagementService, new SyncDescriptor(ExtensionManagementService));
-
-		const instantiationService = new InstantiationService(services);
-		services.set(ILocalizationsService, instantiationService.createInstance(LocalizationsService));
-
-		const extensionManagementCLIService = instantiationService.createInstance(ExtensionManagementCLIService);
-		services.set(IExtensionManagementCLIService, extensionManagementCLIService);
-
-		const ptyService = instantiationService.createInstance(
-			PtyHostService,
-			{
-				graceTime: ProtocolConstants.ReconnectionGraceTime,
-				shortGraceTime: ProtocolConstants.ReconnectionShortGraceTime,
-				scrollback: configurationService.getValue<number>(TerminalSettingId.PersistentSessionScrollback) ?? 100
-			}
-		);
-		services.set(IPtyService, ptyService);
-
-		services.set(IEncryptionMainService, new SyncDescriptor(EncryptionMainService, [machineId]));
-
-		services.set(ICredentialsMainService, new SyncDescriptor(CredentialsMainService, [true]));
-
-		return instantiationService.invokeFunction(accessor => {
-			const remoteExtensionEnvironmentChannel = new RemoteAgentEnvironmentChannel(this._connectionToken, this._environmentService, extensionManagementCLIService, this._logService, this._productService);
-			this._socketServer.registerChannel('remoteextensionsenvironment', remoteExtensionEnvironmentChannel);
-
-			const telemetryChannel = new RemoteTelemetryChannel(accessor.get(IRemoteTelemetryService), appInsightsAppender);
-			this._socketServer.registerChannel('telemetry', telemetryChannel);
-
-			this._socketServer.registerChannel(REMOTE_TERMINAL_CHANNEL_NAME, new RemoteTerminalChannel(this._environmentService, this._logService, ptyService, this._productService));
-
-			const remoteFileSystemChannel = new RemoteAgentFileSystemProviderChannel(this._logService, this._environmentService);
-			this._socketServer.registerChannel(REMOTE_FILE_SYSTEM_CHANNEL_NAME, remoteFileSystemChannel);
-
-			this._socketServer.registerChannel('request', new RequestChannel(accessor.get(IRequestService)));
-
-			const extensionManagementService = accessor.get(IExtensionManagementService);
-			const channel = new ExtensionManagementChannel(extensionManagementService, (ctx: RemoteAgentConnectionContext) => this._getUriTransformer(ctx.remoteAuthority));
-			this._socketServer.registerChannel('extensions', channel);
-
-			const encryptionChannel = ProxyChannel.fromService<RemoteAgentConnectionContext>(accessor.get(IEncryptionMainService));
-			this._socketServer.registerChannel('encryption', encryptionChannel);
-
-			const credentialsChannel = ProxyChannel.fromService<RemoteAgentConnectionContext>(accessor.get(ICredentialsMainService));
-			this._socketServer.registerChannel('credentials', credentialsChannel);
-
-			// clean up deprecated extensions
-			(extensionManagementService as ExtensionManagementService).removeDeprecatedExtensions();
-
-			this._register(new ErrorTelemetry(accessor.get(ITelemetryService)));
-
-			return {
-				telemetryService: accessor.get(ITelemetryService)
-			};
-		});
-	}
-
-	private _getUriTransformer(remoteAuthority: string): IURITransformer {
-		if (!this._uriTransformerCache[remoteAuthority]) {
-			this._uriTransformerCache[remoteAuthority] = createRemoteURITransformer(remoteAuthority);
-		}
-		return this._uriTransformerCache[remoteAuthority];
 	}
 
 	public async handleRequest(req: http.IncomingMessage, res: http.ServerResponse) {
@@ -540,19 +240,6 @@ export class RemoteExtensionHostAgentServer extends Disposable {
 	}
 
 	// Eventually cleanup
-	/**
-	 * Cleans up older logs, while keeping the 10 most recent ones.
-	 */
-	private async _cleanupOlderLogs(logsPath: string): Promise<void> {
-		const currentLog = basename(logsPath);
-		const logsRoot = dirname(logsPath);
-		const children = await Promises.readdir(logsRoot);
-		const allSessions = children.filter(name => /^\d{8}T\d{6}$/.test(name));
-		const oldSessions = allSessions.sort().filter((d) => d !== currentLog);
-		const toDelete = oldSessions.slice(0, Math.max(0, oldSessions.length - 9));
-
-		await Promise.all(toDelete.map(name => Promises.rm(join(logsRoot, name))));
-	}
 
 	private _getRemoteAddress(socket: NodeSocket | WebSocketNodeSocket): string {
 		let _socket: net.Socket;
@@ -965,19 +652,27 @@ export interface IServerAPI {
 }
 
 export async function createServer(address: string | net.AddressInfo | null, args: ServerParsedArgs, REMOTE_DATA_FOLDER: string): Promise<IServerAPI> {
-	const productService = { _serviceBrand: undefined, ...product };
-	const environmentService = new ServerEnvironmentService(args, productService);
+	const connectionToken = parseServerConnectionToken(args);
+	if (connectionToken instanceof ServerConnectionTokenParseError) {
+		console.warn(connectionToken.message);
+		process.exit(1);
+	}
+	const disposables = new DisposableStore();
+	const { socketServer, instantiationService } = await setupServerServices(connectionToken, args, disposables);
 
 	//
 	// On Windows, exit early with warning message to users about potential security issue
 	// if there is node_modules folder under home drive or Users folder.
 	//
-	if (process.platform === 'win32' && process.env.HOMEDRIVE && process.env.HOMEPATH) {
-		const homeDirModulesPath = join(process.env.HOMEDRIVE, 'node_modules');
-		const userDir = dirname(join(process.env.HOMEDRIVE, process.env.HOMEPATH));
-		const userDirModulesPath = join(userDir, 'node_modules');
-		if (fs.existsSync(homeDirModulesPath) || fs.existsSync(userDirModulesPath)) {
-			const message = `
+	instantiationService.invokeFunction((accessor) => {
+		const logService = accessor.get(ILogService);
+
+		if (process.platform === 'win32' && process.env.HOMEDRIVE && process.env.HOMEPATH) {
+			const homeDirModulesPath = join(process.env.HOMEDRIVE, 'node_modules');
+			const userDir = dirname(join(process.env.HOMEDRIVE, process.env.HOMEPATH));
+			const userDirModulesPath = join(userDir, 'node_modules');
+			if (fs.existsSync(homeDirModulesPath) || fs.existsSync(userDirModulesPath)) {
+				const message = `
 
 *
 * !!!! Server terminated due to presence of CVE-2020-1416 !!!!
@@ -990,18 +685,13 @@ export async function createServer(address: string | net.AddressInfo | null, arg
 *
 
 `;
-			const logService = getOrCreateSpdLogService(environmentService);
-			logService.warn(message);
-			console.warn(message);
-			process.exit(0);
+				logService.warn(message);
+				console.warn(message);
+				process.exit(0);
+			}
 		}
-	}
+	});
 
-	const connectionToken = parseServerConnectionToken(args);
-	if (connectionToken instanceof ServerConnectionTokenParseError) {
-		console.warn(connectionToken.message);
-		process.exit(1);
-	}
 	const hasWebClient = fs.existsSync(FileAccess.asFileUri('vs/code/browser/workbench/workbench.html', require).fsPath);
 
 	if (hasWebClient && address && typeof address !== 'string') {
@@ -1010,9 +700,7 @@ export async function createServer(address: string | net.AddressInfo | null, arg
 		console.log(`Web UI available at http://localhost${address.port === 80 ? '' : `:${address.port}`}/${queryPart}`);
 	}
 
-	const remoteExtensionHostAgentServer = new RemoteExtensionHostAgentServer(environmentService, productService, connectionToken, hasWebClient, REMOTE_DATA_FOLDER);
-	const services = await remoteExtensionHostAgentServer.initialize();
-	const { telemetryService } = services;
+	const remoteExtensionHostAgentServer = instantiationService.createInstance(RemoteExtensionHostAgentServer, socketServer, connectionToken, hasWebClient, REMOTE_DATA_FOLDER);
 
 	perf.mark('code/server/ready');
 	const currentTime = performance.now();
@@ -1020,23 +708,27 @@ export async function createServer(address: string | net.AddressInfo | null, arg
 	const vscodeServerListenTime: number = (<any>global).vscodeServerListenTime;
 	const vscodeServerCodeLoadedTime: number = (<any>global).vscodeServerCodeLoadedTime;
 
-	type ServerStartClassification = {
-		startTime: { classification: 'SystemMetaData', purpose: 'PerformanceAndHealth' };
-		startedTime: { classification: 'SystemMetaData', purpose: 'PerformanceAndHealth' };
-		codeLoadedTime: { classification: 'SystemMetaData', purpose: 'PerformanceAndHealth' };
-		readyTime: { classification: 'SystemMetaData', purpose: 'PerformanceAndHealth' };
-	};
-	type ServerStartEvent = {
-		startTime: number;
-		startedTime: number;
-		codeLoadedTime: number;
-		readyTime: number;
-	};
-	telemetryService.publicLog2<ServerStartEvent, ServerStartClassification>('serverStart', {
-		startTime: vscodeServerStartTime,
-		startedTime: vscodeServerListenTime,
-		codeLoadedTime: vscodeServerCodeLoadedTime,
-		readyTime: currentTime
+	instantiationService.invokeFunction((accessor) => {
+		const telemetryService = accessor.get(ITelemetryService);
+
+		type ServerStartClassification = {
+			startTime: { classification: 'SystemMetaData', purpose: 'PerformanceAndHealth' };
+			startedTime: { classification: 'SystemMetaData', purpose: 'PerformanceAndHealth' };
+			codeLoadedTime: { classification: 'SystemMetaData', purpose: 'PerformanceAndHealth' };
+			readyTime: { classification: 'SystemMetaData', purpose: 'PerformanceAndHealth' };
+		};
+		type ServerStartEvent = {
+			startTime: number;
+			startedTime: number;
+			codeLoadedTime: number;
+			readyTime: number;
+		};
+		telemetryService.publicLog2<ServerStartEvent, ServerStartClassification>('serverStart', {
+			startTime: vscodeServerStartTime,
+			startedTime: vscodeServerListenTime,
+			codeLoadedTime: vscodeServerCodeLoadedTime,
+			readyTime: currentTime
+		});
 	});
 
 	if (args['print-startup-performance']) {
@@ -1058,13 +750,3 @@ export async function createServer(address: string | net.AddressInfo | null, arg
 	}
 	return remoteExtensionHostAgentServer;
 }
-
-const getOrCreateSpdLogService: (environmentService: IServerEnvironmentService) => ILogService = (function () {
-	let _logService: ILogService | null;
-	return function getLogService(environmentService: IServerEnvironmentService): ILogService {
-		if (!_logService) {
-			_logService = new LogService(new SpdLogLogger(RemoteExtensionLogFileName, join(environmentService.logsPath, `${RemoteExtensionLogFileName}.log`), true, false, getLogLevel(environmentService)));
-		}
-		return _logService;
-	};
-})();
